Savestate wraparound.
Keep save state index from growing forever when using the auto
increment feature, and add a feature to have reserved indexes
which will not be overwritten by the wraparound.

By selecting reserved index 10, and max_keep as 5, slots 10..15
will be used in rotation, and one gap will always be maintained
due to garbage collection. This gap is used for detecting
last state. Example: slots 10,11,13,14,15 have a file -> gap is 12,
and most recent state is in slot 11.